tests/wpt/web-platform-tests/css/css-writing-modes/parsing/writing-mode-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Writing Modes Level 3: parsing writing-mode with valid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-writing-modes-3/#propdef-writing-mode">
<meta name="assert" content="writing-mode supports the full grammar 'horizontal-tb | vertical-rl | vertical-lr'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("writing-mode", "horizontal-tb");
test_valid_value("writing-mode", "vertical-rl");
test_valid_value("writing-mode", "vertical-lr");
</script>
</body>
</html>
